norm: removed REAL, which was a forbidden constant

+++ include_bonus/utils.h
@@ -15,11 +15,11 @@
 
 typedef void	(*t_objs)(t_scene_obj *obj);
 
-REAL	to_radians(REAL degree);
+double	to_radians(double degree);
 void	calc_components(t_scene *scene);
 void	free_window(t_window *win);
-REAL	real_min(REAL x, REAL y, REAL z);
-REAL	real_max(REAL x, REAL y, REAL z);
+double	real_min(double x, double y, double z);
+double	real_max(double x, double y, double z);
 long	get_current_time(void);
 
 #endif
