+++ srcs_bonus/bvh/add_box.c
@@ -3,7 +3,7 @@
 t_aabb	sphere_aabb(t_sphere *sphere)
 {
 	t_aabb	box;
-	REAL	r;
+	double	r;
 
 	r = sphere->diam * 0.5;
 	box.min = fill_point(sphere->center.x - r, sphere->center.y - r, sphere->center.z - r);
