+++ srcs_bonus/raytracing/phong.c
@@ -19,7 +19,7 @@ t_color	phong_amb_color(t_window *w, t_hit *hit)
 {
 	t_color	color;
 	t_color	amb_color;
-	REAL	light_rate;
+	double	light_rate;
 
 	color.red = 0.0;
 	color.green = 0.0;
@@ -44,8 +44,8 @@ t_color	phong_diffuse_color(t_window *win, t_hit *hit)
 	t_color	diffuse;
 	t_light	light;
 	t_point	light_dir;
-	REAL	diff_factor;
-	REAL	rate;
+	double	diff_factor;
+	double	rate;
 
 	light = *(win->scene_obj->light);
 	light_dir = vec_normalize(vec_sub(light.coord, hit->point));
@@ -94,7 +94,7 @@ int	is_in_shadow(t_window *win, t_hit *hit)
 	t_point	light_dir;
 	t_ray	shadow_ray;
 	t_hit	shadow_hit;
-	REAL	dist_to_light;
+	double	dist_to_light;
 
 	light_dir = vec_normalize(vec_sub(win->scene_obj->light->coord,\
 		hit->point));
@@ -108,4 +108,3 @@ int	is_in_shadow(t_window *win, t_hit *hit)
 	else
 		return (0);
 }
-
